requirements.txt
Django>=5.0
djangorestframework>=3.15
djangorestframework-simplejwt>=5.3
python-dotenv>=1.0
Pillow>=10.3
PyMuPDF==1.26.3
chromadb>=0.5
openai>=1.40.0
tqdm>=4.66.4
